refactor: 使用现代 Rust 特性简化代码
运用 let 链式判断、#[default] 派生宏、is_some_or/is_none_or/is_multiple_of 等新特性重构逻辑，将渲染函数多参数提取为上下文结构体以降低函数签名复杂度，并将常量断言改为编译期检查。

+++ crates/export/src/audio/event.rs
@@ -51,7 +51,7 @@ impl<'a> MidiEventProcessor<'a> {
             tick_conv,
             sink,
             sample_rate: params.sample_rate,
-            channel_count: params.channels.count() as u16,
+            channel_count: params.channels.count(),
             vec_pool: Vec::new(),
         }
     }
